backend/requirements.txt
fastapi
uvicorn
pandas
pandas_ta
numpy
sqlalchemy
pydantic
websockets
python-dotenv
yfinance
psycopg2-binary
feedparser
vaderSentiment
beautifulsoup4
scikit-learn
inngest
httpx
MetaTrader5
alpaca-py
alphalens-reloaded
matplotlib
seaborn
